+++ packages/@ionic/cli/src/lib/integrations/capacitor/utils.ts
@@ -10,3 +10,25 @@ export function generateOptionsForCapacitorBuild(inputs: CommandLineInputs, opti
     platform: platform ? platform : (options['platform'] ? String(options['platform']) : undefined),
   };
 }
+
+export function getNativeIDEForPlatform(platform: string): string {
+  switch (platform) {
+    case 'ios':
+      return 'Xcode';
+    case 'android':
+      return 'Android Studio';
+  }
+
+  return 'Native IDE';
+}
+
+export function getVirtualDeviceNameForPlatform(platform: string): string {
+  switch (platform) {
+    case 'ios':
+      return 'simulator';
+    case 'android':
+      return 'emulator';
+  }
+
+  return 'virtual device';
+}