+++ packages/@ionic/cli/src/lib/project/angular/generate.ts
@@ -100,12 +100,13 @@ To test a generator before file modifications are made, use the ${input('--dry-r
 
   createOptionsFromCommandLine(inputs: CommandLineInputs, options: CommandLineOptions): AngularGenerateOptions {
     const baseOptions = super.createOptionsFromCommandLine(inputs, options);
+    const project = options['project'] ? String(options['project']) : 'app';
 
     // TODO: this is a little gross, is there a better way to pass in all the
     // options that the command got?
     return {
       ...lodash.omit(options, '_', '--', ...GLOBAL_OPTIONS.map(opt => opt.name)),
-      project: options['project'],
+      project,
       ...baseOptions,
     };
   }
@@ -144,9 +145,15 @@ To test a generator before file modifications are made, use the ${input('--dry-r
   }
 
   private async generateComponent(type: string, name: string, options: { [key: string]: string | boolean; }) {
+    const args = {
+      _: ['generate', type, name],
+      // convert `--no-<opt>` style options to `--<opt>=false`
+      ...lodash.mapValues(options, v => v === false ? 'false' : v),
+    };
+
     await this.e.shell.run(
       'ng',
-      unparseArgs({ _: ['generate', type, name], ...options }, {}),
+      unparseArgs(args, { useEquals: true }),
       { cwd: this.e.ctx.execPath, stdio: 'inherit' }
     );
   }